Pipeline fixes: LeRobot version alignment, camera safety, joint validation, training reliability
- Fix LeRobot version mismatch: RunPod now uses same fork as robot (v0.2.0 with ROBOTIS patches)
- Remove dangerous silent camera name remapping in inference (require exact match)
- Guard empty JointTrajectory points (prevent IndexError crash)
- Remove silent joint skip in JointState parsing (fail loudly on missing joints)
- Enforce float32 dtype on state/action arrays in overlay data_manager.py
- Increase RunPod training timeout from 3h to 5h (configurable via timeout_hours)
- Add 3x retry with logging for Supabase progress updates (was silent except:pass)
- Validate steps param required in training API request
- Error-safe cleanup in build script for lerobot copy

+++ physical_ai_tools/physical_ai_server/physical_ai_server/data_processing/data_manager.py
@@ -244,8 +244,8 @@ class DataManager:
         frame = {}
         for camera_name, image in images.items():
             frame[f'observation.images.{camera_name}'] = image
-        frame['observation.state'] = np.array(state)
-        frame['action'] = np.array(action)
+        frame['observation.state'] = np.array(state, dtype=np.float32)
+        frame['action'] = np.array(action, dtype=np.float32)
         self.current_instruction = self._task_info.task_instruction[
             self._current_task % len(self._task_info.task_instruction)
         ]
